chore: dedup toast stub, drop unused tsx dep, fix stale core comment
- Extract the duplicated DOM-toast stub (byte-identical in DocxEditor.tsx and useHyperlinkHandlers.ts) into a shared components/toast.ts.
- Remove the unused root `tsx` devDependency (nothing invokes the tsx CLI; the repo runs .ts via bun). bun install removed it with no dependents.
- Drop the stale '- React wrapper component' line from core's prosemirror/index.ts header (core is React-free; the wrapper lives in the React package).
Zero behavior change (empty changeset).
