+++ package.json
@@ -82,7 +82,6 @@
     "react-dom": "^19.2.7",
     "tinybench": "^6.0.2",
     "tsdown": "0.22.3",
-    "tsx": "^4.22.4",
     "typescript": "6.0.3",
     "use-intl": "^4.13.0"
   },
