+++ packages/react/src/components/toast.ts
@@ -1,0 +1,39 @@
+/**
+ * Toast stub — host apps provide the real toast system; this renders a
+ * temporary DOM banner so the user still sees feedback without a toast provider
+ * (e.g. in the standalone playground). Shared by `DocxEditor` and the hyperlink
+ * handlers so the two never drift.
+ */
+export const toast = (msg: string) => {
+  const existing = document.querySelector("[data-folio-toast]");
+  if (existing) {
+    return;
+  } // debounce rapid calls (e.g., key repeat)
+  const el = document.createElement("div");
+  el.dataset["folioToast"] = "";
+  el.textContent = msg;
+  Object.assign(el.style, {
+    position: "fixed",
+    bottom: "24px",
+    left: "50%",
+    transform: "translateX(-50%)",
+    padding: "10px 20px",
+    borderRadius: "8px",
+    background: "var(--popover, #1f1f1f)",
+    color: "var(--popover-foreground, #fff)",
+    fontSize: "13px",
+    boxShadow: "0 4px 12px var(--doc-shadow-lg, rgba(0,0,0,0.25))",
+    zIndex: "9999",
+    pointerEvents: "none",
+    opacity: "0",
+    transition: "opacity 200ms",
+  });
+  document.body.append(el);
+  requestAnimationFrame(() => {
+    el.style.opacity = "1";
+  });
+  setTimeout(() => {
+    el.style.opacity = "0";
+    setTimeout(() => el.remove(), 200);
+  }, 3000);
+};
